Move tests to MacroScala2Spec
