+++ play-json/shared/src/test/scala/play/api/libs/json/MacroSpec.scala
@@ -560,515 +560,6 @@ class MacroSpec extends AnyWordSpec with Matchers with org.scalatestplus.scalach
         )
       assert(instance.writes(value).value.isInstanceOf[ImmutableLinkedHashMap[?, ?]])
     }
-
-  }
-
-  private def pads(n: Int): JsObject = JsObject(1.to(n).map(i => s"pad$i" -> JsNumber(i)))
-
-  "Reads for classes with more than 22 fields" should {
-    "be generated for simple case class" in {
-      val json     = Json.obj("bar" -> "lorem") ++ pads(22)
-      val expected = BigSimple("lorem", 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22)
-
-      forAll(
-        Json.reads[BigSimple],
-      ) { _.reads(json).mustEqual(JsSuccess(expected)) }
-    }
-
-    "refuse value other than JsObject when properties are optional" in {
-      forAll(Gen.oneOf(Json.reads[BigOptional], Json.format[BigOptional])) { r =>
-        r.reads(Json.obj())
-          .mustEqual(
-            JsSuccess(
-              BigOptional(
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None
-              )
-            )
-          )
-
-        (r.reads(JsString("foo")).asEither match {
-          case Left((_, err :: Nil) :: Nil) =>
-            err.message == "error.expected.jsobject"
-
-          case _ => false
-        }).mustEqual(true)
-      }
-    }
-  }
-
-  "Writes for classes with more than 22 fields" should {
-    "be generated for simple case class" in {
-      Json
-        .writes[BigSimple]
-        .writes(BigSimple("lorem", 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22))
-        .mustEqual(Json.obj("bar" -> "lorem") ++ pads(22))
-    }
-  }
-
-  "Macro for classes with more than 22 fields" should {
-    "handle case class with self type as nested type parameter" when {
-      implicit def eitherReads[A: Reads, B: Reads]: Reads[Either[A, B]]     = TestFormats.eitherReads[A, B]
-      implicit def eitherWrites[A: Writes, B: Writes]: Writes[Either[A, B]] = TestFormats.eitherWrites[A, B]
-      implicit def tuple2Reads[A: Reads, B: Reads]: Reads[(A, B)]           = TestFormats.tuple2Reads[A, B]
-      implicit def tuple2OWrites[A: Writes, B: Writes]: OWrites[(A, B)]     = TestFormats.tuple2OWrites[A, B]
-
-      val jsonNoValue  = Json.obj("id" -> "A") ++ pads(21)
-      val jsonStrValue = Json.obj("id" -> "B", "value" -> "str") ++ pads(21)
-      val jsonFooValue = Json.obj("id" -> "C", "value" -> jsonStrValue) ++ pads(21)
-
-      val fooStrValue =
-        BigFoo(
-          Foo.id("B"),
-          Some(Left("str")),
-          1,
-          2,
-          3,
-          4,
-          5,
-          6,
-          7,
-          8,
-          9,
-          10,
-          11,
-          12,
-          13,
-          14,
-          15,
-          16,
-          17,
-          18,
-          19,
-          20,
-          21
-        )
-      val fooFooValue = BigFoo(
-        Foo.id("C"),
-        Some(Right(fooStrValue)),
-        1,
-        2,
-        3,
-        4,
-        5,
-        6,
-        7,
-        8,
-        9,
-        10,
-        11,
-        12,
-        13,
-        14,
-        15,
-        16,
-        17,
-        18,
-        19,
-        20,
-        21
-      )
-
-      def readSpec(r: Reads[BigFoo]) = {
-        r.reads(jsonNoValue)
-          .mustEqual(
-            JsSuccess(
-              BigFoo(Foo.id("A"), None, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21)
-            )
-          )
-        r.reads(jsonStrValue).mustEqual(JsSuccess(fooStrValue))
-        r.reads(jsonFooValue).mustEqual(JsSuccess(fooFooValue))
-        r.reads(Json.obj("id" -> "D", "value" -> jsonFooValue) ++ pads(21))
-          .mustEqual(
-            JsSuccess(
-              BigFoo(
-                Foo.id("D"),
-                Some(Right(fooFooValue)),
-                1,
-                2,
-                3,
-                4,
-                5,
-                6,
-                7,
-                8,
-                9,
-                10,
-                11,
-                12,
-                13,
-                14,
-                15,
-                16,
-                17,
-                18,
-                19,
-                20,
-                21
-              )
-            )
-          )
-      }
-
-      def writeSpec(w: Writes[BigFoo]) = {
-        w.writes(BigFoo(Foo.id("A"), None, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21))
-          .mustEqual(jsonNoValue)
-        w.writes(fooStrValue).mustEqual(jsonStrValue)
-        w.writes(fooFooValue).mustEqual(jsonFooValue)
-        w.writes(
-          BigFoo(
-            Foo.id("D"),
-            Some(Right(fooFooValue)),
-            1,
-            2,
-            3,
-            4,
-            5,
-            6,
-            7,
-            8,
-            9,
-            10,
-            11,
-            12,
-            13,
-            14,
-            15,
-            16,
-            17,
-            18,
-            19,
-            20,
-            21
-          )
-        ).mustEqual(Json.obj("id" -> "D", "value" -> jsonFooValue) ++ pads(21))
-      }
-
-      "to generate Reads" in readSpec(Json.reads[BigFoo])
-
-      "to generate Writes" in writeSpec(Json.writes[BigFoo])
-
-      "to generate Format" in {
-        val f: OFormat[BigFoo] = Json.format[BigFoo]
-
-        readSpec(f)
-        writeSpec(f)
-      }
-    }
-
-    "handle generic case class with multiple generic parameters" when {
-      val jsonNoOther = Json.obj("base" -> 1) ++ pads(21)
-      val jsonOther   = Json.obj("base" -> 2, "other" -> 3) ++ pads(21)
-
-      val noOther = BigInterval[Int](1, None, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21)
-      val other   =
-        BigInterval[Int](2, Some(3), 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21)
-
-      def readSpec(r: Reads[BigInterval[Int]]) = {
-        r.reads(jsonNoOther).mustEqual(JsSuccess(noOther))
-        r.reads(jsonOther).mustEqual(JsSuccess(other))
-      }
-
-      def writeSpec(r: Writes[BigInterval[Int]]) = {
-        r.writes(noOther).mustEqual(jsonNoOther)
-        r.writes(other).mustEqual(jsonOther)
-      }
-
-      "to generate Reads" in readSpec(Json.reads[BigInterval[Int]])
-
-      "to generate Writes" in writeSpec(Json.writes[BigInterval[Int]])
-
-      "to generate Format" in {
-        val f = Json.format[BigInterval[Int]]
-        readSpec(f)
-        writeSpec(f)
-      }
-    }
-
-    "handle generic case class with multiple generic parameters and self references" when {
-      implicit def eitherReads[A: Reads, B: Reads]: Reads[Either[A, B]]     = TestFormats.eitherReads[A, B]
-      implicit def eitherWrites[A: Writes, B: Writes]: Writes[Either[A, B]] = TestFormats.eitherWrites[A, B]
-      implicit def tuple2Reads[A: Reads, B: Reads]: Reads[(A, B)]           = TestFormats.tuple2Reads[A, B]
-      implicit def tuple2OWrites[A: Writes, B: Writes]: OWrites[(A, B)]     = TestFormats.tuple2OWrites[A, B]
-
-      val nestedLeft  = Json.obj("id" -> 2, "a" -> 0.2F, "b" -> 0.3F, "c" -> 3) ++ pads(20)
-      val nestedRight = Json.obj("id" -> 1, "a" -> 0.1F, "b" -> "right1", "c" -> 2) ++ pads(20)
-
-      val jsonRight = Json.obj(
-        "id" -> 3,
-        "a"  -> nestedRight,
-        "b"  -> "right2",
-        "c"  -> 0.4D
-      ) ++ pads(20)
-
-      val jsonLeft = Json.obj(
-        "id" -> 4,
-        "a"  -> nestedLeft,
-        "b"  -> nestedRight,
-        "c"  -> 0.5D
-      ) ++ pads(20)
-
-      val complexRight = BigComplex(
-        3,
-        BigComplex(1, 0.1F, Right("right1"), 2, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20),
-        Right("right2"),
-        0.4D,
-        1,
-        2,
-        3,
-        4,
-        5,
-        6,
-        7,
-        8,
-        9,
-        10,
-        11,
-        12,
-        13,
-        14,
-        15,
-        16,
-        17,
-        18,
-        19,
-        20
-      )
-
-      val complexLeft = BigComplex(
-        4,
-        BigComplex(2, 0.2F, Left(0.3F), 3, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20),
-        Left(
-          BigComplex(1, 0.1F, Right("right1"), 2, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20)
-        ),
-        0.5D,
-        1,
-        2,
-        3,
-        4,
-        5,
-        6,
-        7,
-        8,
-        9,
-        10,
-        11,
-        12,
-        13,
-        14,
-        15,
-        16,
-        17,
-        18,
-        19,
-        20
-      )
-
-      def readSpec(r: Reads[BigComplex[BigComplex[Float, Int], Double]]) = {
-        r.reads(jsonRight).mustEqual(JsSuccess(complexRight))
-        r.reads(jsonLeft).mustEqual(JsSuccess(complexLeft))
-      }
-
-      def writeSpec(r: Writes[BigComplex[BigComplex[Float, Int], Double]]) = {
-        r.writes(complexRight).mustEqual(jsonRight)
-        r.writes(complexLeft).mustEqual(jsonLeft)
-      }
-
-      "to generate Reads" in readSpec {
-        implicit val nested: Reads[BigComplex[Float, Int]] = Json.reads[BigComplex[Float, Int]]
-        Json.reads[BigComplex[BigComplex[Float, Int], Double]]
-      }
-
-      "to generate Writes" in writeSpec {
-        implicit val nested: OWrites[BigComplex[Float, Int]] = Json.writes[BigComplex[Float, Int]]
-        Json.writes[BigComplex[BigComplex[Float, Int], Double]]
-      }
-
-      "to generate Format" in {
-        implicit val nested: OFormat[BigComplex[Float, Int]] = Json.format[BigComplex[Float, Int]]
-        val f                                                = Json.format[BigComplex[BigComplex[Float, Int], Double]]
-
-        readSpec(f)
-        writeSpec(f)
-      }
-    }
-
-    "handle nesting class" in {
-      implicit val textIdFormat: Format[TextId] = Json.valueFormat[TextId]
-
-      val nesting = new BigNestingClass
-
-      val expected =
-        nesting.Test(
-          Some(new TextId("foo")),
-          1,
-          2,
-          3,
-          4,
-          5,
-          6,
-          7,
-          8,
-          9,
-          10,
-          11,
-          12,
-          13,
-          14,
-          15,
-          16,
-          17,
-          18,
-          19,
-          20,
-          21,
-          22
-        )
-      val expectedJson = Json.obj("underlying" -> "foo") ++ pads(22)
-
-      Json.toJson(expected).mustEqual(expectedJson)
-
-      nesting.Test.format.reads(expectedJson).mustEqual(JsSuccess(expected))
-    }
-
-    "handle case class with generic type and default field" in {
-      implicit val format: Format[BigGenericCaseClassWithDefault[Int]] =
-        Json.format[BigGenericCaseClassWithDefault[Int]]
-
-      val expected =
-        BigGenericCaseClassWithDefault(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 3)
-      val expectedJson = pads(21) ++ Json.obj("data" -> 3, "descr" -> "something")
-
-      Json
-        .toJson(
-          BigGenericCaseClassWithDefault(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 3)
-        )
-        .mustEqual(expectedJson)
-      Json.fromJson(expectedJson).mustEqual(JsSuccess(expected))
-    }
-
-    "handle case class with generic type and overridden default field" in {
-      implicit val format: Format[BigGenericCaseClassWithDefault[Int]] =
-        Json.format[BigGenericCaseClassWithDefault[Int]]
-
-      val expected = BigGenericCaseClassWithDefault(
-        1,
-        2,
-        3,
-        4,
-        5,
-        6,
-        7,
-        8,
-        9,
-        10,
-        11,
-        12,
-        13,
-        14,
-        15,
-        16,
-        17,
-        18,
-        19,
-        20,
-        21,
-        3,
-        "foo"
-      )
-      val expectedJson = pads(21) ++ Json.obj("data" -> 3, "descr" -> "foo")
-
-      Json
-        .toJson(
-          BigGenericCaseClassWithDefault(
-            1,
-            2,
-            3,
-            4,
-            5,
-            6,
-            7,
-            8,
-            9,
-            10,
-            11,
-            12,
-            13,
-            14,
-            15,
-            16,
-            17,
-            18,
-            19,
-            20,
-            21,
-            3,
-            "foo"
-          )
-        )
-        .mustEqual(expectedJson)
-      Json.fromJson(expectedJson).mustEqual(JsSuccess(expected))
-    }
-
-    "field ordering" in {
-      val instance: OWrites[BigFieldOrderTest] = Json.writes[BigFieldOrderTest]
-
-      val value =
-        BigFieldOrderTest(1, 2, 3, Some(4), 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23)
-
-      instance
-        .writes(value)
-        .fields
-        .mustEqual(
-          Seq(
-            "x1"  -> 1,
-            "x2"  -> 2,
-            "x3"  -> 3,
-            "x4"  -> 4,
-            "x5"  -> 5,
-            "x6"  -> 6,
-            "x7"  -> 7,
-            "x8"  -> 8,
-            "x9"  -> 9,
-            "x10" -> 10,
-            "x11" -> 11,
-            "x12" -> 12,
-            "x13" -> 13,
-            "x14" -> 14,
-            "x15" -> 15,
-            "x16" -> 16,
-            "x17" -> 17,
-            "x18" -> 18,
-            "x19" -> 19,
-            "x20" -> 20,
-            "x21" -> 21,
-            "x22" -> 22,
-            "x23" -> 23,
-          ).map { case (k, v) => k -> JsNumber(v) }
-        )
-      assert(instance.writes(value).value.isInstanceOf[ImmutableLinkedHashMap[?, ?]])
-    }
   }
 }
 
@@ -1081,222 +572,6 @@ object MacroSpec {
       x5: Int,
       x6: Int,
   )
-
-  case class BigSimple(
-      bar: String,
-      pad1: Int,
-      pad2: Int,
-      pad3: Int,
-      pad4: Int,
-      pad5: Int,
-      pad6: Int,
-      pad7: Int,
-      pad8: Int,
-      pad9: Int,
-      pad10: Int,
-      pad11: Int,
-      pad12: Int,
-      pad13: Int,
-      pad14: Int,
-      pad15: Int,
-      pad16: Int,
-      pad17: Int,
-      pad18: Int,
-      pad19: Int,
-      pad20: Int,
-      pad21: Int,
-      pad22: Int,
-  )
-
-  case class BigOptional(
-      prop: Option[String],
-      pad1: Option[Int],
-      pad2: Option[Int],
-      pad3: Option[Int],
-      pad4: Option[Int],
-      pad5: Option[Int],
-      pad6: Option[Int],
-      pad7: Option[Int],
-      pad8: Option[Int],
-      pad9: Option[Int],
-      pad10: Option[Int],
-      pad11: Option[Int],
-      pad12: Option[Int],
-      pad13: Option[Int],
-      pad14: Option[Int],
-      pad15: Option[Int],
-      pad16: Option[Int],
-      pad17: Option[Int],
-      pad18: Option[Int],
-      pad19: Option[Int],
-      pad20: Option[Int],
-      pad21: Option[Int],
-      pad22: Option[Int],
-  )
-
-  case class BigFoo(
-      id: Foo.Id,
-      value: Option[Either[String, BigFoo]],
-      pad1: Int,
-      pad2: Int,
-      pad3: Int,
-      pad4: Int,
-      pad5: Int,
-      pad6: Int,
-      pad7: Int,
-      pad8: Int,
-      pad9: Int,
-      pad10: Int,
-      pad11: Int,
-      pad12: Int,
-      pad13: Int,
-      pad14: Int,
-      pad15: Int,
-      pad16: Int,
-      pad17: Int,
-      pad18: Int,
-      pad19: Int,
-      pad20: Int,
-      pad21: Int,
-  )
-
-  case class BigInterval[T](
-      base: T,
-      other: Option[T],
-      pad1: Int,
-      pad2: Int,
-      pad3: Int,
-      pad4: Int,
-      pad5: Int,
-      pad6: Int,
-      pad7: Int,
-      pad8: Int,
-      pad9: Int,
-      pad10: Int,
-      pad11: Int,
-      pad12: Int,
-      pad13: Int,
-      pad14: Int,
-      pad15: Int,
-      pad16: Int,
-      pad17: Int,
-      pad18: Int,
-      pad19: Int,
-      pad20: Int,
-      pad21: Int,
-  )
-
-  case class BigComplex[T, U](
-      id: Int,
-      a: T,
-      b: Either[T, String],
-      c: U,
-      pad1: Int,
-      pad2: Int,
-      pad3: Int,
-      pad4: Int,
-      pad5: Int,
-      pad6: Int,
-      pad7: Int,
-      pad8: Int,
-      pad9: Int,
-      pad10: Int,
-      pad11: Int,
-      pad12: Int,
-      pad13: Int,
-      pad14: Int,
-      pad15: Int,
-      pad16: Int,
-      pad17: Int,
-      pad18: Int,
-      pad19: Int,
-      pad20: Int,
-  )
-
-  case class BigGenericCaseClassWithDefault[A](
-      pad1: Int,
-      pad2: Int,
-      pad3: Int,
-      pad4: Int,
-      pad5: Int,
-      pad6: Int,
-      pad7: Int,
-      pad8: Int,
-      pad9: Int,
-      pad10: Int,
-      pad11: Int,
-      pad12: Int,
-      pad13: Int,
-      pad14: Int,
-      pad15: Int,
-      pad16: Int,
-      pad17: Int,
-      pad18: Int,
-      pad19: Int,
-      pad20: Int,
-      pad21: Int,
-      data: A,
-      descr: String = "something",
-  )
-
-  case class BigFieldOrderTest(
-      x1: Int,
-      x2: Int,
-      x3: Int,
-      x4: Option[Int],
-      x5: Int,
-      x6: Int,
-      x7: Int,
-      x8: Int,
-      x9: Int,
-      x10: Int,
-      x11: Int,
-      x12: Int,
-      x13: Int,
-      x14: Int,
-      x15: Int,
-      x16: Int,
-      x17: Int,
-      x18: Int,
-      x19: Int,
-      x20: Int,
-      x21: Int,
-      x22: Int,
-      x23: Int,
-  )
-
-  class BigNestingClass {
-    case class Test(
-        underlying: Option[TextId],
-        pad1: Int,
-        pad2: Int,
-        pad3: Int,
-        pad4: Int,
-        pad5: Int,
-        pad6: Int,
-        pad7: Int,
-        pad8: Int,
-        pad9: Int,
-        pad10: Int,
-        pad11: Int,
-        pad12: Int,
-        pad13: Int,
-        pad14: Int,
-        pad15: Int,
-        pad16: Int,
-        pad17: Int,
-        pad18: Int,
-        pad19: Int,
-        pad20: Int,
-        pad21: Int,
-        pad22: Int,
-    )
-
-    object Test {
-      implicit def format(implicit textIdFormat: Format[TextId]): Format[Test] =
-        Json.format[Test]
-    }
-  }
 
   sealed trait Family
   case class Simple(bar: String)            extends Family
